+++ src/coreclr/tools/Common/Compiler/DisplayNameHelpers.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Reflection.Metadata;
 using System.Text;
 
 using Internal.TypeSystem;
-
+using Internal.TypeSystem.Ecma;
 using Debug = System.Diagnostics.Debug;
 
 namespace ILCompiler
@@ -38,6 +39,12 @@ namespace ILCompiler
             {
                 sb.Append(method.OwningType.GetDisplayNameWithoutNamespace());
             }
+            else if (method.GetPropertyForAccessor() is PropertyPseudoDesc property)
+            {
+                sb.Append(property.Name);
+                sb.Append('.');
+                sb.Append(property.GetMethod.Name == method.Name ? "get" : "set");
+            }
             else
             {
                 sb.Append(method.Name);
@@ -66,6 +73,20 @@ namespace ILCompiler
             sb.Append(')');
 
             return sb.ToString();
+        }
+
+        public static string GetParameterDisplayName(this EcmaMethod method, int parameterIndex)
+        {
+            var reader = method.MetadataReader;
+            var methodDefinition = reader.GetMethodDefinition(method.Handle);
+            foreach (var parameterHandle in methodDefinition.GetParameters())
+            {
+                var parameter = reader.GetParameter(parameterHandle);
+                if (parameter.SequenceNumber == parameterIndex + 1)
+                    return reader.GetString(parameter.Name);
+            }
+
+            return $"#{parameterIndex}";
         }
 
         public static string GetDisplayName(this FieldDesc field)
